refactor: move vertex code to an embed module
this will also be the module that holds R/W code for generic attributes
